+++ internal/api/handlers/user.go
@@ -89,7 +89,7 @@ func (h *UserHandler) List(w http.ResponseWriter, r *http.Request) {
 	limit := pb.GetInt64("limit", 50, 1, 100)
 	offset := pb.GetInt64("offset", 0, 0, 100000)
 
-	users, err := h.services.User.List(ctx, limit, offset)
+	users, err := h.services.User.List(ctx, int32(limit), int32(offset))
 	if err != nil {
 		writeServiceError(w, h.logger, &reqID, "list users", err)
 		return
